+++ internal/kms/kms.go
@@ -139,10 +139,6 @@ func (k *Kms) GetExternalWrappers() *ExternalWrappers {
 	return ret
 }
 
-func generateKeyId(scopeId string, purpose KeyPurpose, version uint32) string {
-	return fmt.Sprintf("%s_%s_%d", scopeId, purpose, version)
-}
-
 // GetWrapper returns a wrapper for the given scope and purpose. When a keyId is
 // passed, it will ensure that the returning wrapper has that key ID in the
 // multiwrapper. This is not necesary for encryption but should be supplied for
@@ -168,8 +164,11 @@ func (k *Kms) GetWrapper(ctx context.Context, scopeId string, purpose KeyPurpose
 	val, ok := k.scopePurposeCache.Load(scopeId + purpose.String())
 	if ok {
 		wrapper := val.(*multiwrapper.MultiWrapper)
-		if opts.withKeyId == "" || wrapper.WrapperForKeyID(opts.withKeyId) != nil {
+		if opts.withKeyId == "" {
 			return wrapper, nil
+		}
+		if keyIdWrapper := wrapper.WrapperForKeyID(opts.withKeyId); keyIdWrapper != nil {
+			return keyIdWrapper, nil
 		}
 		// Fall through to refresh our multiwrapper for this scope/purpose from the DB
 	}
@@ -191,6 +190,13 @@ func (k *Kms) GetWrapper(ctx context.Context, scopeId string, purpose KeyPurpose
 		return nil, errors.Wrap(err, op, errors.WithMsg(fmt.Sprintf("error loading %s for scope %s", purpose.String(), scopeId)))
 	}
 	k.scopePurposeCache.Store(scopeId+purpose.String(), wrapper)
+
+	if opts.withKeyId != "" {
+		if keyIdWrapper := wrapper.WrapperForKeyID(opts.withKeyId); keyIdWrapper != nil {
+			return keyIdWrapper, nil
+		}
+		return nil, errors.New(errors.KeyNotFound, op, "unable to find specified key ID")
+	}
 
 	return wrapper, nil
 }
